requirements.txt
reportlab
beautifulsoup4
requests
pillow
python-dotenv
